refactor: split the oracle review stage into its own module
The gate-size extractions pushed _review_oracle_gates.py to 82% of the code cap and gave _apply_post_review_stages a ninth argument. The peer-review stage moves to a sibling alongside the red-team one, and the four red-team wiring inputs travel as one RedTeamStageConfig.
