+++ src/synthorg/engine/_review_oracle_gates.py
@@ -13,9 +13,7 @@ from typing import TYPE_CHECKING, NamedTuple
 from synthorg.core.task import Task
 from synthorg.core.task_enums import (
     BlockedReason,
-    Stakes,
     TaskStatus,
-    compare_stakes,
 )
 from synthorg.engine.completion_oracle.evaluator import BuildTestOracle
 from synthorg.engine.completion_oracle.protocol import CompletionOracleGate
@@ -24,13 +22,11 @@ from synthorg.engine.completion_oracle.review_models import (
     CompletionOracleGateResult,
     CompletionOracleVerdict,
 )
-from synthorg.engine.review_gate_inputs import DeliverableReviewInputBuilder
 from synthorg.observability import get_logger
 from synthorg.observability.events.approval_gate import APPROVAL_GATE_REVIEW_REWORK
 from synthorg.observability.events.completion_oracle import (
     BUILD_TEST_GATE_BLOCKED,
     COMPLETION_ORACLE_ESCALATION_ROUTED,
-    COMPLETION_ORACLE_GATE_SKIPPED,
     COMPLETION_ORACLE_REWORK_ROUTED,
     COMPLETION_ORACLE_SHADOW_OBSERVED,
 )
@@ -309,175 +305,3 @@ def _route_non_approving_verdict(
         event=APPROVAL_GATE_REVIEW_REWORK,
         approved=False,
     )
-
-
-def _resolve_oracle_activation(
-    task: Task,
-    *,
-    gate: CompletionOracleGate | None,
-    min_stakes: Stakes,
-) -> tuple[bool, bool]:
-    """Decide whether the judge runs on this task, and say why when it does not.
-
-    Only THIS gate's own escalation is answered by the human's decision.
-    Keyed on the recorded reason, never on the status: BLOCKED is reached
-    from several directions (a coordination wave releasing a subtask is one,
-    and an unstaffed reviewer role is another), and a status-only check
-    silently exempted those from the verification this gate exists to impose.
-    ``REVIEWER_UNSTAFFED`` deliberately does NOT qualify: nobody was ever
-    asked, so there is no decision to preserve and the judge must run again.
-
-    Args:
-        task: The task whose completion is being judged.
-        gate: The peer-review gate, when one is wired.
-        min_stakes: The stakes floor the gate is armed at.
-
-    Returns:
-        ``(oracle_active, judge_already_ruled)``.
-    """
-    judge_already_ruled = (
-        task.status is TaskStatus.BLOCKED
-        and task.blocked_reason is BlockedReason.ORACLE_ESCALATED
-    )
-    if judge_already_ruled:
-        # Re-running the judge on the human's answer re-escalates, which parks
-        # the task again: the decision the escalation exists to obtain is
-        # discarded by the rule that requested it. Whether a human is needed
-        # and what the human decides are two separately owned questions; this
-        # returns the second to its owner.
-        #
-        # Only the judge is skipped. The deliverable is still built by the
-        # caller, because the red-team gate and the output-policy backstop are
-        # different authorities that have not ruled on anything, and handing
-        # them ``None`` reads as "retrieval failed": the red-team gate then
-        # fails closed and reroutes the approval it was never asked about.
-        logger.info(
-            COMPLETION_ORACLE_GATE_SKIPPED,
-            task_id=str(task.id),
-            reason="human_decision_owns_an_escalated_task",
-            note="task parked by an earlier escalation; the decision is the answer",
-        )
-    oracle_active = (
-        gate is not None
-        and not judge_already_ruled
-        and compare_stakes(task.stakes, min_stakes) >= 0
-    )
-    return oracle_active, judge_already_ruled
-
-
-def _no_deliverable_outcome(task: Task) -> GateOutcome:
-    """Block completion when an enforced oracle has nothing to inspect.
-
-    Fail CLOSED on enforcement mode, not builder presence: an enforced oracle
-    that cannot obtain a reviewable deliverable, whether the builder returned
-    ``None`` OR none is wired at all, must not let the task reach COMPLETED
-    unreviewed. Shadow mode only observes, so it never reaches here and
-    preserves the incoming outcome instead.
-
-    Args:
-        task: The task being judged, for the log.
-
-    Returns:
-        The rework outcome that keeps the task short of COMPLETED.
-    """
-    logger.warning(
-        COMPLETION_ORACLE_GATE_SKIPPED,
-        task_id=str(task.id),
-        reason="no_deliverable_block",
-        note=(
-            "Completion oracle is active but no reviewable deliverable was "
-            "retrievable; blocking completion (fail-closed)."
-        ),
-    )
-    return GateOutcome(
-        target=TaskStatus.IN_PROGRESS,
-        transition_reason=(
-            "Completion review could not retrieve a deliverable to inspect."
-        ),
-        event=APPROVAL_GATE_REVIEW_REWORK,
-        approved=False,
-    )
-
-
-async def apply_oracle_review_stage(
-    *,
-    completion_oracle_gate: CompletionOracleGate | None,
-    completion_oracle_shadow_mode: bool,
-    completion_oracle_min_stakes: Stakes,
-    deliverable_input_builder: DeliverableReviewInputBuilder | None,
-    red_team_active: bool,
-    output_policy_active: bool,
-    task: Task,
-    outcome: GateOutcome,
-) -> tuple[GateOutcome, RedTeamReviewInput | None]:
-    """Run the peer-review gate and hand back the shared deliverable input.
-
-    Resolves the reviewable deliverable ONCE (shared with the downstream
-    red-team gate and the output-policy backstop, so a completion where several
-    consumers are active pays a single retrieval) whenever the oracle is active
-    at this task's stakes, the red-team gate will consume it, or the
-    output-policy backstop is enabled (the last is stakes-independent, so a
-    low-stakes deliverable is still policy-checked). An ENFORCED (non-shadow)
-    oracle fails CLOSED whenever no deliverable is retrievable -- whether the
-    builder returned ``None`` or none is wired -- because the peer-review gate
-    would otherwise receive a ``None`` input and silently preserve approval,
-    letting the task reach COMPLETED without the independent review the oracle
-    promises. Shadow mode only observes, so it never blocks. Then applies the
-    stakes-gated peer-review gate.
-
-    Returns:
-        The (possibly rerouted) ``(target, reason, event, approved)`` tuple and
-        the built deliverable input (``None`` when no consumer needed it), so
-        the caller's red-team gate and output-policy backstop can reuse it
-        without a second retrieval.
-    """
-    target, transition_reason, event, approved, _ = outcome
-    oracle_active, judge_already_ruled = _resolve_oracle_activation(
-        task,
-        gate=completion_oracle_gate,
-        min_stakes=completion_oracle_min_stakes,
-    )
-    deliverable_input = (
-        await deliverable_input_builder.build(task)
-        if deliverable_input_builder is not None
-        and (oracle_active or red_team_active or output_policy_active)
-        else None
-    )
-    if (
-        oracle_active
-        and deliverable_input is None
-        and not completion_oracle_shadow_mode
-    ):
-        return _no_deliverable_outcome(task), deliverable_input
-    if completion_oracle_gate is None:
-        return (
-            GateOutcome(target, transition_reason, event, approved),
-            deliverable_input,
-        )
-    if not oracle_active:
-        # The escalated case already logged its own reason above; saying
-        # "below_stakes_threshold" here as well would name a cause that is
-        # not the one that applied.
-        if not judge_already_ruled:
-            logger.info(
-                COMPLETION_ORACLE_GATE_SKIPPED,
-                task_id=str(task.id),
-                reason="below_stakes_threshold",
-                stakes=task.stakes.value,
-                min_stakes=completion_oracle_min_stakes.value,
-            )
-        return (
-            GateOutcome(target, transition_reason, event, approved),
-            deliverable_input,
-        )
-    outcome = await apply_completion_oracle_gate(
-        gate=completion_oracle_gate,
-        review_input=to_oracle_input(deliverable_input, task),
-        shadow_mode=completion_oracle_shadow_mode,
-        task_id=str(task.id),
-        target=target,
-        transition_reason=transition_reason,
-        event=event,
-        approved=approved,
-    )
-    return outcome, deliverable_input
